:recycle: refactor: getSharedQuestions 관련 네이밍 복수형으로 통일 #37
